Moved to utils
